fix: address nits
