+++ crates/near-mpc-crypto-types/src/ckd.rs
@@ -111,3 +111,25 @@ impl CKDRequest {
         }
     }
 }
+
+#[cfg(test)]
+#[expect(non_snake_case)]
+mod tests {
+    use super::*;
+    use crate::{Bls12381G1PublicKey, CKDAppPublicKey, CKDRequest};
+    use near_account_id::AccountId;
+
+    #[test]
+    fn ckd_request_new__should_derives_app_id_deterministically() {
+        let account_id: AccountId = "alice.near".parse().unwrap();
+        let pk = CKDAppPublicKey::AppPublicKey(Bls12381G1PublicKey([1u8; 48]));
+        let domain_id = DomainId(0);
+
+        let r1 = CKDRequest::new(pk.clone(), domain_id, &account_id, "path/a");
+        let r2 = CKDRequest::new(pk.clone(), domain_id, &account_id, "path/a");
+        assert_eq!(r1.app_id, r2.app_id);
+
+        let r3 = CKDRequest::new(pk, domain_id, &account_id, "path/b");
+        assert_ne!(r1.app_id, r3.app_id);
+    }
+}
